# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "vault-server",
    "description": "Arguflow REST API OpenAPI Documentation",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/topic": {
      "get": {
        "tags": [
          "topic"
        ],
        "operationId": "get_all_topics",
        "responses": {
          "200": {
            "description": "All topics belonging to a given user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Topic"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Service error relating to topic get",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DefaultError"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "topic"
        ],
        "operationId": "create_topic",
        "requestBody": {
          "description": "JSON request payload to create chat topic",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTopicData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The JSON response payload containing the created topic",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Topic"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Topic resolution empty or a service error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DefaultError"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "topic"
        ],
        "operationId": "update_topic",
        "requestBody": {
          "description": "JSON request payload to update a chat topic",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTopicData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Confirmation that the topic was updated"
          },
          "400": {
            "description": "Service error relating to topic update",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DefaultError"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "topic"
        ],
        "operationId": "delete_topic",
        "requestBody": {
          "description": "JSON request payload to delete a chat topic",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteTopicData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Confirmation that the topic was deleted"
          },
          "400": {
            "description": "Service error relating to topic deletion",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DefaultError"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTopicData": {
        "type": "object",
        "required": [
          "resolution"
        ],
        "properties": {
          "normal_chat": {
            "type": "boolean",
            "nullable": true
          },
          "resolution": {
            "type": "string"
          }
        }
      },
      "DefaultError": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "DeleteTopicData": {
        "type": "object",
        "required": [
          "topic_id"
        ],
        "properties": {
          "topic_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "Topic": {
        "type": "object",
        "required": [
          "id",
          "user_id",
          "resolution",
          "side",
          "deleted",
          "created_at",
          "updated_at",
          "normal_chat"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "deleted": {
            "type": "boolean"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "normal_chat": {
            "type": "boolean"
          },
          "resolution": {
            "type": "string"
          },
          "side": {
            "type": "boolean"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "UpdateTopicData": {
        "type": "object",
        "required": [
          "topic_id",
          "resolution",
          "side"
        ],
        "properties": {
          "resolution": {
            "type": "string"
          },
          "side": {
            "type": "boolean"
          },
          "topic_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    }
  },
...
}